requirements.txt
accelerate
aioboto3
aiofiles
aiohttp
asyncpg
click
# database packages
graspologic
gremlinpython
hnswlib
nano-vectordb
neo4j
networkx

numpy
ollama
openai
oracledb
psycopg-pool
psycopg[binary,pool]
pydantic
pymilvus
pymongo
pymysql
python-dotenv
pyvis
setuptools
# lmdeploy[all]
sqlalchemy
streamlit
streamlit-agraph
tenacity


# LLM packages
tiktoken
torch
tqdm
transformers
xxhash

lightrag-hku